Update KPM.jl
This would be my proposed API. I also tweaked the docstring, in particular removing the requirement that ket is a *sparse* matrix. We should make this generic, in case the needed ket is actually not sparse.

This change appears as a pull request to your pull request. Simple merge it on your fork to accept it.